Clean up old files & examples
