+++ RELEASE.md
@@ -1,7 +1,14 @@
 # Release Notes
 
 ## Version 0.1.0, ~
+* Battery dispatch and simulation with example in examples/simulate_hybrid_wbattery_dispatch.py 
 * Separate layout-related functions from optimization code into Layout classes
+* Refactor Optimizer classes to use HybridSimulation with examples in examples/optimization/hybrid_npv.py and examples/optimization/hybrid_sizing_problem.py
+* Add Floris as custom module for Wind simulations with examples in examples/add_custom_module
+* Move plotting functions into dedicated files
+* Rename "Solar" classes to "PV"
+* Add ElectricityPrices class with some example input files in resource_files/grid
+* Add storage costs to CostCalculator
 
 ## Version 0.0.5, Apr 30, 2021
 * Update PySAM requirements
