+++ examples/optimization/hybrid_sizing_problem.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from collections import OrderedDict, namedtuple
 from hybrid.sites import make_circular_site, make_irregular_site, SiteInfo, locations
 from hybrid.hybrid_simulation import HybridSimulation
-from tools.optimization.optimization_problem_new import OptimizationProblem
+from tools.optimization.optimization_problem import OptimizationProblem
 
 
 site = 'irregular'
