+++ hybrid/keys.py
@@ -17,7 +17,10 @@ def get_developer_nrel_gov_key():
     return developer_nrel_gov_key
 
 
-def set_nrel_key_dot_env():
-    load_dotenv()
+def set_nrel_key_dot_env(path=None):
+    if path:
+        load_dotenv(path)
+    else:
+        load_dotenv()
     NREL_API_KEY = os.getenv("NREL_API_KEY")
     set_developer_nrel_gov_key(NREL_API_KEY)
